support forfile run and correct overlap

+++ python/sfllm/layers/sampler.py
@@ -52,7 +52,7 @@ class Sampler(nn.Module):
 
         sampled_index = torch.multinomial(probs_sort, num_samples=1)
         # int32 range is enough to represent the token ids
-        probs_idx = probs_idx.to(torch.int32)
+        # probs_idx = probs_idx.to(torch.int32)
         batch_next_token_ids = torch.gather(probs_idx, dim=1, index=sampled_index).view(-1)
         return batch_next_token_ids
